First release for DHCP Lease Calculator: v0.1
This tool is to provide an estimate on the Leases Per Seconds based on the number of clients, lease time and expected growth values. The calculated LPS is a best guessed value to assist in sizing.
NOTE: It is best to capture the accurate value from the existing production environment to determine the required capacity.